refactor: move to useCarouselEvents
