+++ components/carousel/CarouselTypeLatestSales.vue
@@ -9,7 +9,7 @@
 </template>
 
 <script lang="ts" setup>
-import { useCarouselNftEvents } from './utils/useCarousel'
+import { useCarouselNftEvents } from './utils/useCarouselEvents'
 
 const { urlPrefix } = usePrefix()
 const { nfts, ids } = useCarouselNftEvents({ type: 'latestSales' })
